added editor object to manage the imgui render and editor informations

+++ Engine/src/data/Mesh.cpp
@@ -50,7 +50,7 @@ void Mesh::Draw(Shader& shader)
 
 int Mesh::GetNumberOfTriangles() const
 {
-   return Vertices.size() / 3;
+   return (int)Vertices.size() / 3;
 }
 
 #pragma endregion
